\name{wheatarea}
\alias{wheatarea}
\docType{data}
\title{ Wheat area measurement and satellite data.}
\description{ 
Wheat area data measured at the scene in the block of Yanzhou District, Jining City, Shandong Province. The data corresponding to each block comes from the ArcGIS platform. The whole dataset consists of a total number of 458 villages and 14750 wheat blocks.

 }
\usage{data(wheatarea)}
\format{
  A data frame with 14708 observations on the following 3 variables.
  \describe{
    \item{\code{pixel}:}{ Pixel sizes of each wheat blocks.}
    \item{\code{F_AREA}:}{ Field inspection area of each wheat blocks.}
    \item{\code{code}:}{ Street code.}
    }
}

\source{ 
- Liu Y, Qu W, Cui Z, Liu X, Xu W, Jiang j,. (2021). Estimation of Wheat Growing Area via Mixed Model Prediction Using Satellite Data. Journal of Applied Statistics and Management. 
}

\keyword{datasets}
